Create ContextMenu project
Context menu installer project and tests project. Registers shell extension for Windows Explorer with a command to start Cloney.Wizard and passes in the filepath to the chosen Visual Studio solution file. The menu option can be accessed by right-clicking on a solution file. Can also unregister the shell extension.